codigo atualizado

+++ IFitness/src/main/java/br/edu/ifsp/arq/tsi/arqweb2/ifitness/model/Activity.java
@@ -12,7 +12,7 @@ public class Activity implements Serializable {
 	private ActivityType type;
 	private LocalDate date;
 	private Double distance;
-	private Double duration;
+	private Integer duration;
 	private User user;
 
 	public Long getId() {
@@ -47,11 +47,11 @@ public class Activity implements Serializable {
 		this.distance = distance;
 	}
 
-	public Double getDuration() {
+	public Integer getDuration() {
 		return duration;
 	}
 
-	public void setDuration(Double duration) {
+	public void setDuration(Integer duration) {
 		this.duration = duration;
 	}
